pyproject.toml
[tool.poetry]
name = "flux-trainer"
version = "0.1.0"
description = "A UI application for training and using FLUX image generation models"
authors = []
readme = "README.md"
packages = [{include = "flux_trainer"}]

[tool.poetry.dependencies]
python = "^3.9"
gradio = "^4.0.0"
requests = "^2.31.0"
fire = "^0.5.0"
python-dotenv = "^1.0.0"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 